Refactor browse_foreigners functions to static methods
